device_simulator/requirements.txt
paho-mqtt>=2.2.1
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0
